Fix some problems and implementing stream instead for

+++ src/main/java/br/com/movies/dao/MoviesDAO.java
@@ -1,11 +1,13 @@
 package br.com.movies.dao;
 
+import java.math.BigInteger;
+
 import br.com.hibernate.GenericDAO;
 import br.com.movies.entity.MoviesEntity;
 
 public interface MoviesDAO extends GenericDAO<MoviesEntity, Integer>{
 	
 //	verify if genres already were persisteds.
-	Integer veriFyGenreData();
+	BigInteger veriFyGenreData();
 
 }
